Move tupleselect
