+++ src/components/TupleSelectWithRegexConditional/TupleSelectWithRegexConditional.stories.tsx
@@ -1,21 +1,20 @@
 import { Disclaimer } from "@leafygreen-ui/typography";
 import { action } from "@storybook/addon-actions";
-import { ProjectFilterOptions } from "types/commits";
 import TupleSelectWithRegexConditional from ".";
 
 export default {
-  title: "Pages/Commits/TupleSelectWithRegexConditional",
+  title: "Components/TupleSelect/TupleSelectWithRegexConditional",
   component: TupleSelectWithRegexConditional,
 };
 
 const options = [
   {
-    value: ProjectFilterOptions.BuildVariant,
+    value: "build_variant",
     displayName: "Build Variant",
     placeHolderText: "Search Build Variant names",
   },
   {
-    value: ProjectFilterOptions.Task,
+    value: "task",
     displayName: "Task",
     placeHolderText: "Search Task names",
   },
